Feat: add category to main page
